feat: delete asset_downloader service and add cache
