+++ kili/queries/asset/exceptions.py
@@ -1,5 +1,5 @@
 """
-Asset download service exceptions
+Asset queries related exceptions
 """
