+++ kili/services/__init__.py
@@ -2,9 +2,8 @@
 Python SDK service layer
 """
 
-from kili.services.asset_download import download_asset_media
 from kili.services.asset_import import import_assets
 from kili.services.export import export_labels
 from kili.services.label_import import import_labels_from_files
 
-__all__ = ["export_labels", "import_assets", "import_labels_from_files", "download_asset_media"]
+__all__ = ["export_labels", "import_assets", "import_labels_from_files"]
